add admin area, make User entity work with both Asp.net identity va Repository

+++ src/Core/HvCommerce.Core/ApplicationServices/HvRoleStore.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 
 namespace HvCommerce.Core.ApplicationServices
 {
-    public class HvRoleStore : RoleStore<Role, UserRole, RoleClaim, HvDbContext, int>
+    public class HvRoleStore : RoleStore<Role, UserRole, RoleClaim, HvDbContext, long>
     {
         public HvRoleStore(HvDbContext context)
         : base(context)
